Fixed Gayness
Uhm... so I don't know, but this was an issue for me where the item wasn't being taken from the inventory if it's been named on the anvil to something different. 

also... why the fuck would you want an item to pop out on the ground where some asshole can come by and take it ? 
I made it so you receive the item in your inventory.